apps/orchestrator/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "orchestrator"
version = "0.1.0"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "httpx",
    "pydantic-settings",
    "orjson",
    "langchain",
    "langgraph",
    "langchain-community",
]
